scripts/requirements.txt
# 資料匯入腳本依賴套件（MOE + TEPA）
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
openpyxl>=3.1.0